requirements.txt
PyQt6>=6.6.0
pyqtgraph>=0.13.4
matplotlib>=3.8.0
pyaudio>=0.2.14
websockets>=12.0
httpx>=0.27.0
pymupdf>=1.24.0
python-docx>=1.1.0
spacy>=3.7.0
sentence-transformers>=2.7.0
librosa>=0.10.2
scikit-learn>=1.5.0
numpy>=1.26.0
sounddevice>=0.4.7
groq>=0.9.0
elevenlabs>=1.3.0
python-dotenv>=1.0.0
reportlab>=4.2.0
scipy>=1.13.0
